Split get_service function into module
Avoid having two files with this same function.  That way enhancements to the function can be done only once.
